+++ src/ipstack/tls_common.c
@@ -97,7 +97,7 @@ static err_t tls_client_recv(void *arg, struct altcp_pcb *pcb, struct pbuf *p, e
         pbuf_copy_partial(p, buf, p->tot_len, 0);
         buf[p->tot_len] = 0;
 
-        printf("***\nnew data received from server:\n***\n\n%s\n", buf);
+        printf("***\nnew data received from server:\n***\n\n[%s]\n", buf);
         free(buf);
 
         altcp_recved(pcb, p->tot_len);
