+++ src/painless/PressureSensor.cpp
@@ -11,7 +11,7 @@ PressureSensor::PressureSensor(std::shared_ptr<PicoI2C> i2c, uint16_t device_add
 }
 
 int PressureSensor::readPressure() {
-    i2c->transaction(address, data, 1, values, 2);
+    pressure_i2c->transaction(address, data, 1, values, 2);
     pressure = ( (values[0] << 8) | values[1]) /240 *0.95;  // 0.95 is based on sea level ( based on datasheet)
     pressure = (pressure <= 0) ? 0 : (pressure >= 127) ? 0 : pressure;
     return pressure;
