+++ src/painless/co2valve.cpp
@@ -4,7 +4,7 @@
 
 #include "co2valve.h"
 
-co2valve::co2valve(int valve_pin) {
+co2valve::co2valve(int valve_pin): valve_pin(valve_pin){
     gpio_init(valve_pin);
     gpio_set_dir(valve_pin, GPIO_OUT);
     gpio_pull_up(valve_pin);
@@ -12,18 +12,18 @@ co2valve::co2valve(int valve_pin) {
 }
 
 void co2valve::open() {
-    gpio_put(valve, 1);
+    gpio_put(valve_pin, 1);
 }
 
 void co2valve::close() {
-    gpio_put(valve, 0);
+    gpio_put(valve_pin, 0);
 }
 
 void co2valve::setValve(int value) {
-    valve = value;
+   // valve = value;
 }
 
 int co2valve::getValve() {
-    return valve;
+    //return valve;
 }
